+++ starkbank/__init__.py
@@ -1,4 +1,4 @@
-version = "2.23.0"
+version = "2.24.0"
 
 user = None
 language = "en-US"